Remove reference to java.xml.ws.annotation.
This is not needed any more, because the code was upgraded to a newer
version from upstream which migrated away from using
javax.annotation.Generated (see https://r.android.com/1099815). The
redundant --add-modules option breaks builds with -source 9 -target 9
(so the Java module system is enabled) using OpenJDK 11 (which removes
this module). In other words, the third of the 'Test' lines below
fails without this change and succeeds with it.

Test: make javaparser
Test: EXPERIMENTAL_JAVA_LANGUAGE_LEVEL_9=true make javaparser
Test: EXPERIMENTAL_JAVA_LANGUAGE_LEVEL_9=true EXPERIMENTAL_USE_OPENJDK11_TOOLCHAIN=true make javaparser
Test: . $ANDROID_BUILD_TOP/frameworks/base/tests/Codegen/runTest.sh (in downstream branch)
Bug: 131683177
Change-Id: Id0e6adcb4ba56b0d15f904891c5fda655953f4f3