Adapt NodeList.getFirst/getLast to standard method
These names are reserved by List<T> to return objets with a type of T
now. The methods that used to return Optional<T> are removed and the
default implementation is used instead.

Tracking changes in various files to either check size() first if the
callsite was using ifPresent() pattern, or to directly get the object if
the callsite was simply calling .get() on the result.

Flag: EXEMPT library import upgrade
Bug: 370425985
Test: m javaparser
Test: m codegen_cli
Test: atest protologtool-tests
Change-Id: I6616a31849ab99fc94158669ac083afbcfb93031